Creacion de mis productos y cambios en lista de productos
Se realizo la creacion del componente de mis productos y se reestructuro el componente de la lista de productos generales
